Feature/display usage stats improved dx

+++ src/build-datadir-image.sh
@@ -12,9 +12,17 @@ error() {
     exit $?
 }
 
+show_system_state() {
+  echo "Dumping usage status:"
+  echo "w"
+  w
+  echo "df -h"
+  df -h
+}
+
 # Remove all temporary data we can get our hands on.
 cleanup() {
-  if [[ ! -z "${TMP_DATADIR-}" ]] 
+  if [[ ! -z "${TMP_DATADIR-}" ]]
     then
     echo "delete datadir"
     rm -rf "${TMP_DATADIR}"
@@ -70,11 +78,11 @@ INITSCRIPT=""
 # Let the user specify a init-script to be run after the db-import.
 if [ $# -gt 2 ]
   then
-    if [[ $3 == standard-* ]] 
+    if [[ $3 == standard-* ]]
       then
       INITSCRIPT="${SCRIPTDIR}/init-sql/${3:9}/reset.sql"
     else
-      echo "init script must be one of standard-drupal7 or standard-drupal8"
+      echo "ERROR: init script must be one of standard-drupal7 or standard-drupal8"
       exit 1
     fi
   else
@@ -82,6 +90,8 @@ if [ $# -gt 2 ]
 fi
 
 $SCRIPTDIR/init-gcloud.sh
+
+show_system_state
 
 # Get the tag and make sure it is available as a dbdump image.
 docker pull "${DUMP_IMAGE_SOURCE}"
@@ -117,6 +127,8 @@ docker container create \
 # Now that the container has been created (but not yet started) we can copy
 # files to its volumes.
 
+show_system_state
+
 # Pick up a init-script (ie. sql that should be run after the databasedump has
 # been imported) specified on the commandline, and add it to docker-compose.yml.
 if [ ! -z $INITSCRIPT ]
@@ -135,20 +147,21 @@ docker start -a "${DB_CONTAINER_NAME}"
 TMP_DATADIR=$(mktemp -d --suffix=datadir)
 docker cp -a "${DB_CONTAINER_NAME}:/var/lib/mysql" "${TMP_DATADIR}/mysql"
 
+show_system_state
+
 # Build the pre-init data-container, use same tag as the sql-dump image.
 echo "Building the datadir image ${DATADIR_IMAGE_DESTINATION}"
 
 # Ok ... so ....  aufs (which will host the datadir for fulldb build) has this
 # "feature" where if you create a file/directory with a given ownership and
-# permission, subsequent permissions can only be narrower.
-# In other words, if we add, say, a datadir owned by root and then change the
-# ownership to mysql - strange things can happen when you then try to change
-# things. Specifically 
-#   touch datadir/db/a && rm datadir/db/a 
-# will fail on the rm. So, to avoid all of that, we do something slightly crazy
-# , we make everything in the datadir world write/readable - make a mental note
-# of never ever allowing this image to hit prod - and then add the datadir. 
-# This way we start out with very liberal permissions which makes aufs happy.
+# permission, subsequent permissions can only be narrower. In other words, if we
+# add, say, a datadir owned by root and then change the ownership to mysql -
+# strange things can happen when you then try to change things. Specifically
+# touch datadir/db/a && rm datadir/db/a will fail on the rm.
+# So, to avoid all of that, we do something slightly crazy, we make everything
+# in the datadir world write/readable - make a mental note of never ever
+# allowing this image to hit prod - and then add the datadir. This way we start
+# out with very liberal permissions which makes aufs happy.
 # This might be resolved in a future version of Docker so it is worth checking
 # whether the following can be removed when this script is updated.
 chmod -R a+rw "${TMP_DATADIR}"
@@ -156,6 +169,15 @@ find "${TMP_DATADIR}" -type d -print0 | xargs -0 chmod a+x
 
 # Build using same tag as the one from dbdump.
 docker build --tag "${DATADIR_IMAGE_DESTINATION}" -f "Dockerfile" "${TMP_DATADIR}"
-echo "Pushing ${DATADIR_IMAGE_DESTINATION}"
-docker push "${DATADIR_IMAGE_DESTINATION}"
-docker rmi "${DATADIR_IMAGE_DESTINATION}"
+
+show_system_state
+
+if [[ -z "${NO_PUSH-}" ]]; then
+  echo "Pushing ${DATADIR_IMAGE_DESTINATION}"
+  docker push "${DATADIR_IMAGE_DESTINATION}"
+  docker rmi "${DATADIR_IMAGE_DESTINATION}"
+else
+  echo "Datadir image is available as ${DATADIR_IMAGE_DESTINATION}"
+fi
+
+show_system_state
